arch: move rv32i code to general riscv crate
This code will support both 64 bit and 32 bit platforms, so making it
general (even when it now only supports 32 bit platforms) helps with
future changes.

Assembly is hard to review when it has moved and changed. This isolates
the just the move step.

The only meaningful change I had to make was adding:

```
// Mock implementation for tests on Travis-CI.
pub extern "C" fn _start() {
    unimplemented!()
}
```

to make `cargo check` work.
